migrate proto structs to proto.go files
