--- external/proto/status/service.proto
@@ -1,11 +1,0 @@
-syntax = "proto2";
-
-// Request is what the Status service is expected to receive from clients.
-message Request {
-}
-
-// Response is what the Status service will reply to clients.
-message Response {
-  optional Status status = 1;
-  optional ServerIdentity serveridentity = 2;
-}
