+++ pop/service/struct.go
@@ -71,68 +71,6 @@ const (
 	PopStatusOK
 )
 
-// CheckConfig asks whether the pop-config and the attendees are available.
-type CheckConfig struct {
-	PopHash   []byte
-	Attendees []kyber.Point
-}
-
-// CheckConfigReply sends back an integer for the Pop. 0 means no config yet,
-// other values are defined as constants.
-// If PopStatus == PopStatusOK, then the Attendees will be the common attendees between
-// the two nodes.
-type CheckConfigReply struct {
-	PopStatus int
-	PopHash   []byte
-	Attendees []kyber.Point
-}
-
-// MergeConfig asks if party is ready to merge
-type MergeConfig struct {
-	// FinalStatement of current party
-	Final *FinalStatement
-	// Hash of PopDesc party to merge with
-	ID []byte
-}
-
-// MergeConfigReply responds with info of asked party
-type MergeConfigReply struct {
-	// status of merging process
-	PopStatus int
-	// hash of party was asking to merge
-	PopHash []byte
-	// FinalStatement of party was asked to merge
-	Final *FinalStatement
-}
-
-// PinRequest will print a random pin on stdout if the pin is empty. If
-// the pin is given and is equal to the random pin chosen before, the
-// public-key is stored as a reference to the allowed client.
-type PinRequest struct {
-	Pin    string
-	Public kyber.Point
-}
-
-// StoreConfig presents a config to store
-type StoreConfig struct {
-	Desc      *PopDesc
-	Signature []byte
-}
-
-// StoreConfigReply gives back the hash.
-// TODO: StoreConfigReply will give in a later version a handler that can be used to
-// identify that config.
-type StoreConfigReply struct {
-	ID []byte
-}
-
-// FinalizeRequest asks to finalize on the given descid-popconfig.
-type FinalizeRequest struct {
-	DescID    []byte
-	Attendees []kyber.Point
-	Signature []byte
-}
-
 func (fr *FinalizeRequest) hash() ([]byte, error) {
 	h := cothority.Suite.Hash()
 	_, err := h.Write(fr.DescID)
@@ -152,42 +90,3 @@ func (fr *FinalizeRequest) hash() ([]byte, error) {
 	return h.Sum(nil), nil
 }
 
-// FinalizeResponse returns the FinalStatement if all conodes already received
-// a PopDesc and signed off. The FinalStatement holds the updated PopDesc, the
-// pruned attendees-public-key-list and the collective signature.
-type FinalizeResponse struct {
-	Final *FinalStatement
-}
-
-// FetchRequest asks to get FinalStatement
-type FetchRequest struct {
-	ID               []byte
-	ReturnUncomplete *bool
-}
-
-// MergeRequest asks to start merging process for given Party
-type MergeRequest struct {
-	ID        []byte
-	Signature []byte
-}
-
-// GetProposals asks the conode to return a list of all waiting proposals. A waiting
-// proposal is either deleted after 1h or if it has been confirmed using
-// StoreConfig.
-type GetProposals struct {
-}
-
-// GetProposalsReply returns the list of all waiting proposals on that node.
-type GetProposalsReply struct {
-	Proposals []PopDesc
-}
-
-// VerifyLink returns if a given public key is linked.
-type VerifyLink struct {
-	Public kyber.Point
-}
-
-// VerifyLinkReply returns true if the public key is in the admin-list.
-type VerifyLinkReply struct {
-	Exists bool
-}
