+++ skipchain/msgs.go
@@ -62,98 +62,6 @@ func init() {
 // This file holds all messages that can be sent to the SkipChain,
 // both from the outside and between instances of this service
 
-// External calls
-
-// StoreSkipBlock - Requests a new skipblock to be appended to the given
-// SkipBlock. If the given TargetSkipChainID is an empty slice, then a genesis
-// block is created.  Otherwise, the new block is added to the skipchain
-// specified by TargetSkipChainID.
-type StoreSkipBlock struct {
-	TargetSkipChainID SkipBlockID
-	NewBlock          *SkipBlock
-	Signature         *[]byte
-}
-
-// StoreSkipBlockReply - returns the signed SkipBlock with updated backlinks
-type StoreSkipBlockReply struct {
-	Previous *SkipBlock
-	Latest   *SkipBlock
-}
-
-// GetUpdateChain - the client sends the hash of the last known
-// Skipblock and will get back a list of all necessary SkipBlocks
-// to get to the latest.
-type GetUpdateChain struct {
-	LatestID SkipBlockID
-}
-
-// GetUpdateChainReply - returns the shortest chain to the current SkipBlock,
-// starting from the SkipBlock the client sent
-type GetUpdateChainReply struct {
-	Update []*SkipBlock
-}
-
-// GetAllSkipchains - returns all known last blocks of skipchains.
-type GetAllSkipchains struct {
-}
-
-// GetAllSkipchainsReply - returns all known last blocks of skipchains.
-type GetAllSkipchainsReply struct {
-	SkipChains []*SkipBlock
-}
-
-// Internal calls
-
-// PropagateSkipBlocks sends a newly signed SkipBlock to all members of
-// the Cothority
-type PropagateSkipBlocks struct {
-	SkipBlocks []*SkipBlock
-}
-
-// ForwardSignature is called once a new skipblock has been accepted by
-// signing the forward-link, and then the older skipblocks need to
-// update their forward-links. Each cothority needs to get the necessary
-// blocks and propagate the skipblocks itself.
-type ForwardSignature struct {
-	// TargetHeight is the index in the backlink-slice of the skipblock
-	// to update
-	TargetHeight int
-	// Previous is the second-newest skipblock
-	Previous SkipBlockID
-	// Newest is the newest skipblock, signed by previous
-	Newest *SkipBlock
-}
-
-// GetSingleBlock asks for a single block.
-type GetSingleBlock struct {
-	ID SkipBlockID
-}
-
-// GetSingleBlockByIndex asks for a single block at a certain index. If Index == -1,
-// the last block on the skipchain is returned.
-type GetSingleBlockByIndex struct {
-	Genesis SkipBlockID
-	Index   int
-}
-
-// Internal calls
-
-// GetBlock asks for an updated block, in case for a conode that is not
-// in the roster-list of that block.
-type GetBlock struct {
-	ID SkipBlockID
-}
-
-// PropagateSkipBlock sends a newly signed SkipBlock to all members of
-// the Cothority
-type PropagateSkipBlock struct {
-	SkipBlock *SkipBlock
-}
-
-// GetBlockReply returns the requested block.
-type GetBlockReply struct {
-	SkipBlock *SkipBlock
-}
 
 // Protocol messages
